fix: add .toDouble() to avoid type error in _pow function
math.pow() returns num, but the function return type is double

+++ lib/models/models.dart
@@ -218,7 +218,7 @@ class PlayerProfileModel {
   }
   
   static double _pow(int base, double exponent) {
-    return math.pow(base, exponent);
+    return math.pow(base, exponent).toDouble();
   }
   
   // Get XP required for next level
